refactor(handlers): move handlers from their own package into api
